Assignment3/Python-URL_Shortner/requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==22.0.4
pymongo==4.3.3
setuptools==58.1.0
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
mysql-connector-python==8.0.33